Move notify action to reducer
