+++ assets/js/state/notifications.js
@@ -1,1 +1,6 @@
 export const NOTIFICATION = 'NOTIFICATION';
+
+export const notify = ({ text, icon }) => ({
+  type: NOTIFICATION,
+  payload: { text, icon },
+});
